requirements.txt
Pillow
PyPDF2
python-docx
tqdm
piexif
